ai-voice-assistant/requirements.txt
# AI Voice Assistant — Dependencies
# Core Framework
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0

# Audio Processing
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.0

# Speech Recognition
openai-whisper==20231117
openai==1.35.3

# Utilities
python-dotenv==1.0.1
